extract CanReceiver to new file imported only by CAN BMS
